solução de pequenos bugs
git-svn-id: svn://localhost/barragem/trunk@202 a01e8a6f-d01e-c94b-8b3f-b6c26fcad43f

+++ src/net/barragem/view/backingbean/GerirJogadorBean.java
@@ -24,7 +24,7 @@ public class GerirJogadorBean extends BaseBean {
 	private String jogadorNome;
 	private String pesquisa;
 	private String pesquisaSalva;
-	private Integer tipoPesquisa = new Integer(2);
+	private Integer tipoPesquisa = new Integer(1);
 	private ListDataModel jogadores;
 	private Paginavel<Jogador> paginacaoJogadores;
